+++ src/main/java/com/github/zastrixarundell/torambot/Values.java
@@ -1,7 +1,12 @@
 package com.github.zastrixarundell.torambot;
 
+import java.awt.image.BufferedImage;
+import java.util.Properties;
+
 public class Values
 {
+
+    private static String version = "";
 
     public final static String footerMessage = "Support me by going on: http://corneey.com/w2ObhY";
 
@@ -10,6 +15,8 @@ public class Values
     public final static String inviteLink = "http://ceesty.com/w2Ncfe";
 
     public final static String donationLink = "https://donatebot.io/checkout/602112468961067011";
+
+    public static BufferedImage dyeImage = null;
 
     private static boolean ranOnHostingService = false;
 
@@ -23,4 +30,24 @@ public class Values
 
     public static String getPrefix() { return prefix; }
 
+    public static BufferedImage getDyeImage() { return dyeImage; }
+
+    public static void setDyeImage(BufferedImage dyeImage) { Values.dyeImage = dyeImage; }
+
+    public static void getMavenVersion()
+    {
+        try
+        {
+            final Properties properties = new Properties();
+            properties.load(Values.class.getClassLoader().getResourceAsStream("values.properties"));
+            version = properties.getProperty("version");
+        }
+        catch (Exception ignore)
+        {
+
+        }
+
+    }
+
+    public static String getVersion() { return version; }
 }