+++ src/main/java/com/github/zastrixarundell/torambot/commands/corynwebsite/ItemCommand.java
@@ -2,6 +2,7 @@ package com.github.zastrixarundell.torambot.commands.corynwebsite;
 
 import com.github.zastrixarundell.torambot.Parser;
 import com.github.zastrixarundell.torambot.Values;
+import com.github.zastrixarundell.torambot.objects.Item;
 import org.javacord.api.entity.message.embed.EmbedBuilder;
 import org.javacord.api.event.message.MessageCreateEvent;
 import org.javacord.api.listener.message.MessageCreateListener;
@@ -12,7 +13,7 @@ import org.jsoup.select.Elements;
 
 import java.util.ArrayList;
 
-public class Item implements MessageCreateListener
+public class ItemCommand implements MessageCreateListener
 {
 
     @Override
@@ -37,7 +38,8 @@ public class Item implements MessageCreateListener
 
         String data = String.join(" ", arguments);
 
-        Runnable runnable = () ->
+        Runnable runnable;
+        runnable = () ->
         {
             try
             {
@@ -56,24 +58,24 @@ public class Item implements MessageCreateListener
         (new Thread(runnable)).start();
     }
 
-    private ArrayList<com.github.zastrixarundell.torambot.objects.Item> getItems(Element body)
+    private ArrayList<Item> getItems(Element body)
     {
 
         Elements trs = body.getElementsByTag("tr");
 
-        ArrayList<com.github.zastrixarundell.torambot.objects.Item> listOfItems = new ArrayList<>();
+        ArrayList<Item> listOfItems = new ArrayList<>();
 
         for(int size = 0, count = 0; size < trs.size() && count < 5; size++)
             if(trs.get(size).parent() == body)
             {
-                listOfItems.add(new com.github.zastrixarundell.torambot.objects.Item(trs.get(size)));
+                listOfItems.add(new Item(trs.get(size)));
                 count++;
             }
 
         return listOfItems;
     }
 
-    private void sendItemEmbed(com.github.zastrixarundell.torambot.objects.Item item, MessageCreateEvent messageCreateEvent)
+    private void sendItemEmbed(Item item, MessageCreateEvent messageCreateEvent)
     {
         EmbedBuilder embed = new EmbedBuilder()
                 .setTitle(item.getName())