+++ src/main/java/com/github/zastrixarundell/torambot/commands/toramwebsite/LatestCommand.java
@@ -10,7 +10,7 @@ import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
-public class Maintenance implements MessageCreateListener
+public class LatestCommand implements MessageCreateListener
 {
 
     @Override
@@ -20,15 +20,14 @@ public class Maintenance implements MessageCreateListener
         if (!messageCreateEvent.getMessageAuthor().isRegularUser())
             return;
 
-        if (!messageCreateEvent.getMessageContent().toLowerCase().startsWith(Values.getPrefix() + "maintenance"))
-            if(!messageCreateEvent.getMessageContent().toLowerCase().startsWith(Values.getPrefix() + "maint"))
-                return;
+        if (!messageCreateEvent.getMessageContent().toLowerCase().startsWith(Values.getPrefix() + "latest"))
+            return;
 
         Runnable runnable = () ->
         {
             try
             {
-                Document document = Jsoup.connect("https://en.toram.jp/?type_code=update#contentArea").get();
+                Document document = Jsoup.connect("https://en.toram.jp/?type_code=all#contentArea").get();
                 Elements links = document.select("a[href]");
 
                 for (Element link : links)
@@ -65,14 +64,13 @@ public class Maintenance implements MessageCreateListener
                 }
 
             }
-            catch (Exception exception)
+            catch (Exception ignore)
             {
-                Methods.sendErrorMessage(messageCreateEvent);
+                MethodsCommand.sendErrorMessage(messageCreateEvent);
             }
         };
 
         (new Thread(runnable)).start();
-
     }
 
 }