+++ src/main/java/com/github/zastrixarundell/torambot/commands/toramwebsite/NewsCommand.java
@@ -10,7 +10,7 @@ import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 
-public class News implements MessageCreateListener
+public class NewsCommand implements MessageCreateListener
 {
 
     @Override
@@ -62,7 +62,7 @@ public class News implements MessageCreateListener
             }
             catch (Exception exception)
             {
-                Methods.sendErrorMessage(messageCreateEvent);
+                MethodsCommand.sendErrorMessage(messageCreateEvent);
             }
         };